Resolve "Enhance streaming graph"

Closes #161

See merge request !465

+++ compss/runtime/adaptors/gat/worker/src/main/java/es/bsc/compss/gat/worker/implementations/BinaryDefinition.java
@@ -75,8 +75,8 @@ public class BinaryDefinition extends ImplementationDefinition {
     }
 
     @Override
-    public int getTimeOut() {
-        return 0;
+    public long getTimeOut() {
+        return 0L;
     }
 
 }
