+++ compss/runtime/adaptors/gat/worker/src/main/java/es/bsc/compss/gat/worker/implementations/OMPSsDefinition.java
@@ -75,8 +75,8 @@ public class OMPSsDefinition extends ImplementationDefinition {
     }
 
     @Override
-    public int getTimeOut() {
-        return 0;
+    public long getTimeOut() {
+        return 0L;
     }
 
 }
